refactor(866): carve Humans.Monitor out of AuditLog (G5)
AuditLogController had three actions that were monitoring rather than audit
browsing, and two of them injected GoogleIntegration services. AuditLog is a
horizontal, so peters-hard-rules.md forbids that reach — it was invisible only
because both ends were still Base-resident. It becomes an assembly-level
violation the moment GoogleIntegration goes to G5, caught by
AuditLogArchitectureTests.SectionReferencesNoVerticalSection.

DriveActivityMonitorService turned out to be the same shape one level down: it
injects IGoogleDriveActivityClient, ITeamResourceService, ISystemSettingsService,
IUserServiceRead and IAuditLogService, and calls no repository — a cross-section
orchestrator by the hard rules' own definition, sitting in
Services/GoogleIntegration/ on code locality alone.

Monitor may reference both GoogleIntegration and AuditLog because Monitor is not
a horizontal: it is a leaf consumer, and its whole outward surface is
Humans.Monitor.Contracts (one interface, one method, returning int) for
DriveActivityMonitorJob, which stays in Base per G5 template step 6b.

Moved: DriveActivityMonitorService, IDriveActivityMonitorService, GoogleSync.cshtml,
the GoogleSyncAudit view models, and the CheckDriveActivity / Resource / Human
actions. AuditLog keeps the general audit browser, which reaches no section.

Routes move to /Monitor/*; the three call sites (the audit page's toolbar form,
Google's sync tab, the human admin detail page) are repointed. The
EndpointAuthorizationTests policy rows are retargeted rather than dropped — the
policies are unchanged and a deleted row would read as "no longer critical".

Table-less section: no DbContext, no G4 gate, and the tests/Directory.Build.props
EF exclusion that shape needs.

Build clean, 0 warnings. Full suite green; Humans.Integration.Tests 243/244
(1 pre-existing skip) run alone — the two Store order failures under
full-solution concurrency are the known flake ledger entry, not this change.

Refs nobodies-collective/Humans#866
