+++ docs/sections/_Index.md
@@ -20,9 +20,10 @@ move recipe is [`G5-SECTION-TEMPLATE.md`](G5-SECTION-TEMPLATE.md).
 
 | Section | Project | Invariants doc |
 |---|---|---|
-| Audit Log | `src/Sections/Humans.AuditLog` | [AuditLog.md](../../src/Sections/Humans.AuditLog/Docs/AuditLog.md) — the append path and the `/AuditLog` pages; the name-resolving read path (`AuditViewerService`, `AuditEvent`) stays in `Humans.Application`, because it injects Users', Teams' and GoogleIntegration's read interfaces and a *horizontal* section may not reference a vertical |
+| Audit Log | `src/Sections/Humans.AuditLog` | [AuditLog.md](../../src/Sections/Humans.AuditLog/Docs/AuditLog.md) — the append path and the `/AuditLog` pages; the name-resolving read path (`AuditViewerService`, `AuditEvent`) stays in `Humans.Application`, because it injects Users', Teams' and GoogleIntegration's read interfaces and a *horizontal* section may not reference a vertical; the three Google-sync monitoring actions moved to `Humans.Monitor` for the same reason, one level up |
 | Auth | `src/Sections/Humans.Auth` | [Auth.md](../../src/Sections/Humans.Auth/Docs/Auth.md) — `role_assignments`, the role-assignment service and its §15 decorator; the magic-link sign-in path (`MagicLinkService`, `AccountController`, `Views/Account/*`) stays in Base, because that service is an orchestrator injecting `Humans.Email.Contracts` and a *horizontal* section may not reference a vertical |
 | Agent | `src/Sections/Humans.Agent` | [Agent.md](../../src/Sections/Humans.Agent/Docs/Agent.md) |
+| Monitor | `src/Sections/Humans.Monitor` | [Monitor.md](../../src/Sections/Humans.Monitor/Docs/Monitor.md) — Drive-activity anomaly detection and the Google-sync audit trail; carved out of AuditLog because a *horizontal* may not reference a vertical, and `DriveActivityMonitorService` injects five sections while calling no repository |
 | Calendar | `src/Sections/Humans.Calendar` | [Calendar.md](../../src/Sections/Humans.Calendar/Docs/Calendar.md) |
 | Campaigns | `src/Sections/Humans.Campaigns` | [Campaigns.md](../../src/Sections/Humans.Campaigns/Docs/Campaigns.md) |
 | Cantina | `src/Sections/Humans.Cantina` | [Cantina.md](../../src/Sections/Humans.Cantina/Docs/Cantina.md) |
@@ -89,7 +90,7 @@ Cross-check against [`design-rules.md` §8 (Table Ownership Map)](../architectur
 | **Holded** | `HoldedController` (`Humans.Holded.Controllers`) — `/Holded` admin screen + `/Holded/Accounts/{number}` GL page | — | `Service` (`Humans.Holded.Services`) | `Repository` / `IHoldedMirrorRepository` (`Humans.Holded.Data`) | `holded_ledger_lines`, `holded_accounts`, `holded_api_calls`, `holded_sync_states` |
 | **Gate** | `GateController`, `GateVendorBackfillAdminController` (`Humans.Gate.Controllers`) | — | `GateService` (`Humans.Gate.Services`) | `GateRepository` / `IGateRepository` (`Humans.Gate.Data`) | `gate_scan_events`, `gate_settings`, `gate_staff_pins` |
 | **Governance** | `GovernanceController`, `GovernanceApplicationsController`, `GovernanceBoardVotingController` (`Humans.Governance.Controllers`) | `GovernanceIndexService`, `MembershipCalculator` (`Humans.Governance.Services`) | `ApplicationDecisionService` (`Humans.Governance.Services`) | `ApplicationRepository` / `IApplicationRepository` (`Humans.Governance.Data`) | `applications`, `application_state_history`, `board_votes` |
-| **Google Integration** | `GoogleController` | `GoogleGroupSyncService`, `GoogleAdminService`, `EmailProvisioningService`, `GoogleRemovalNotificationService` | `GoogleWorkspaceSyncService`, `GoogleWorkspaceUserService`, `DriveActivityMonitorService`, `SyncSettingsService`, `TeamResourceService`, Google clients (`GoogleDirectoryClient`, `GoogleGroupMembershipClient`, `GoogleGroupProvisioningClient`, `GoogleDriveActivityClient`, `GoogleDrivePermissionsClient`, `WorkspaceUserDirectoryClient`) | `GoogleResourceRepository`, `GoogleSyncOutboxRepository`, `DriveActivityMonitorRepository`, `SyncSettingsRepository` | `google_resources`, `google_sync_outbox`, `sync_service_settings`, `system_settings` (key `DriveActivityMonitor:LastRunAt`) |
+| **Google Integration** | `GoogleController` | `GoogleGroupSyncService`, `GoogleAdminService`, `EmailProvisioningService`, `GoogleRemovalNotificationService` | `GoogleWorkspaceSyncService`, `GoogleWorkspaceUserService`, `SyncSettingsService`, `TeamResourceService`, Google clients (`GoogleDirectoryClient`, `GoogleGroupMembershipClient`, `GoogleGroupProvisioningClient`, `GoogleDriveActivityClient`, `GoogleDrivePermissionsClient`, `WorkspaceUserDirectoryClient`) | `GoogleResourceRepository`, `GoogleSyncOutboxRepository`, `DriveActivityMonitorRepository`, `SyncSettingsRepository` | `google_resources`, `google_sync_outbox`, `sync_service_settings`, `system_settings` (key `DriveActivityMonitor:LastRunAt`) |
 | **Guide** | `GuideController` (`Humans.Guide.Controllers`) | — | `GuideContentService`, `GuideRoleResolver`, `GuideRenderer` (`Humans.Guide.Services`); `GitHubGuideContentSource` stays in `Humans.Infrastructure` — a shared GitHub-markdown fetcher, three of whose four consumers are not Guide's | — | — (content served from GitHub `docs/guide/`, cached via `IMemoryCache`) |
 | **Holded Connector** | — (no UI) | — | `HoldedClient`, `HoldedCallLog` (`Humans.Infrastructure.Services.Holded`) | — | — (thin API client in Base; owns no tables — see [`Holded.md`](Holded.md)) |
 | **Issues** | `IssuesController`, `IssuesApiController` (`Humans.Issues.Controllers`) | — | `IssuesService` (`Humans.Issues.Services`) | `IssuesRepository` / `IIssuesRepository` (`Humans.Issues.Data`) | `issues`, `issue_comments` |
@@ -110,6 +111,7 @@ Cross-check against [`design-rules.md` §8 (Table Ownership Map)](../architectur
 | **Search** | `SearchController` (`Humans.Search.Controllers`, internal) | `SearchService` (`Humans.Search.Services`, internal) | — | — | — (owns no tables; fans out to Users, Teams, Camps, Shifts and Events through their service interfaces) |
 | **Mailer** | `MailerAdminController` (`Humans.Mailer.Controllers`) | `MailerImportService`, `MailerAudienceSyncService` | `MailerLiteClient` (`Humans.Mailer.Services.MailerLite`) | — | — (MailerLite is the system of record; in-Humans writes route through other sections' services) |
 | **Scanner** | `ScannerController` (`src/Sections/Humans.Scanner`) | — | — | — | — (presentational; owns no tables) |
+| **Monitor** | `MonitorController` (`Humans.Monitor.Controllers`, internal) | `DriveActivityMonitorService` (`Humans.Monitor.Services`, internal) | — | — | — (owns no tables; reads Google through GoogleIntegration's connector, writes through `IAuditLogService`) |
 | **Tour** | `TourController` (`src/Sections/Humans.Tour`) | — | — | — | — (presentational; owns no tables) |
 | **Debug** | `DebugController` (`Humans.Debug.Controllers`), `LogApiController`, `ColorPaletteController`, `WidgetGalleryController`, `TimezoneApiController` | — | — | `AdminDatabaseDiagnosticsRepository` | — (Debug owns no tables; it reads in-memory trackers and `IAdminDatabaseDiagnosticsService`) |
 | **Development** | `DevLoginController`, `DevSeedController` (`Humans.Development.Controllers`) | — | `DevPersonaSeeder`, `DevelopmentCampRoleSeeder`, `DevelopmentDashboardSeeder` (`Humans.Development.Services`) — dev fixture seeders, not application services; registered outside Production only | — | — (dev-only tooling; owns no tables and writes only through other sections' services) |
