+++ src/Humans.Infrastructure/Humans.Infrastructure.csproj
@@ -46,6 +46,7 @@
     <ProjectReference Include="..\Sections\Humans.Onboarding.Contracts\Humans.Onboarding.Contracts.csproj" />
     <ProjectReference Include="..\Sections\Humans.Surveys.Contracts\Humans.Surveys.Contracts.csproj" />
     <ProjectReference Include="..\Sections\Humans.Teams.Contracts\Humans.Teams.Contracts.csproj" />
+    <ProjectReference Include="..\Sections\Humans.Monitor.Contracts\Humans.Monitor.Contracts.csproj" />
     <!--
       Tickets' contracts leaf: AdminDatabaseDiagnosticsService reads the order projection,
       TicketSyncJob drives ITicketSync, and GateVendorCheckInJob mirrors a gate admission
