+++ src/Humans.Web/Extensions/Sections/GoogleIntegrationSectionExtensions.cs
@@ -5,7 +5,6 @@ using GoogleSyncSettingsService = Humans.Application.Services.GoogleIntegration.
 using GoogleSyncOutboxService = Humans.Application.Services.GoogleIntegration.GoogleSyncOutboxService;
 using GoogleEmailProvisioningService = Humans.Application.Services.GoogleIntegration.EmailProvisioningService;
 using GoogleAdminService = Humans.Application.Services.GoogleIntegration.GoogleAdminService;
-using GoogleDriveActivityMonitorService = Humans.Application.Services.GoogleIntegration.DriveActivityMonitorService;
 using GoogleRemovalNotificationService = Humans.Application.Services.GoogleIntegration.GoogleRemovalNotificationService;
 
 namespace Humans.Web.Extensions.Sections;
@@ -20,7 +19,6 @@ internal static class GoogleIntegrationSectionExtensions
         services.AddSingleton<IGoogleResourceRepository, GoogleResourceRepository>();
         services.AddSingleton<IGoogleSyncOutboxRepository, GoogleSyncOutboxRepository>();
         services.AddScoped<IGoogleSyncOutboxService, GoogleSyncOutboxService>();
-        services.AddScoped<IDriveActivityMonitorService, GoogleDriveActivityMonitorService>();
         services.AddScoped<IGoogleAdminService, GoogleAdminService>();
         services.AddScoped<IGoogleRemovalNotificationService, GoogleRemovalNotificationService>();
