+++ src/Humans.Web/Views/UsersAdmin/AdminDetail.cshtml
@@ -323,7 +323,7 @@
                         </button>
                     </form>
                 }
-                <a asp-controller="AuditLog" asp-action="Human" asp-route-id="@Model.UserId"
+                <a asp-controller="Monitor" asp-action="Human" asp-route-id="@Model.UserId"
                    class="btn btn-outline-secondary btn-sm w-100">View Google Sync Audit</a>
             </div>
         </div>
