+++ src/Sections/Humans.AuditLog/Views/AuditLog/_Content.cshtml
@@ -4,7 +4,7 @@
 @model Humans.AuditLog.Models.AuditLogListViewModel
 
 <page-header title="Audit Log">
-    <form asp-controller="AuditLog" asp-action="CheckDriveActivity" method="post" class="d-inline" id="checkForm">
+    <form asp-controller="Monitor" asp-action="CheckDriveActivity" method="post" class="d-inline" id="checkForm">
         @Html.AntiForgeryToken()
         <button type="submit" class="btn btn-outline-warning" id="checkBtn">
             Check Drive Activity Now
