+++ src/Sections/Humans.Monitor/Humans.Monitor.csproj
@@ -1,0 +1,62 @@
+<Project Sdk="Microsoft.NET.Sdk.Razor">
+
+  <PropertyGroup>
+    <!-- MVC views/controllers, not Blazor components. -->
+    <AddRazorSupportForMvc>true</AddRazorSupportForMvc>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <InternalsVisibleTo Include="Humans.Monitor.Tests" />
+    <!-- The one assembly allowed to see many sections (design §5). -->
+    <InternalsVisibleTo Include="Humans.Integration.Tests" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <!--
+      Microsoft.NET.Sdk.Web adds these to ImplicitUsings; Microsoft.NET.Sdk.Razor does not.
+      Declaring them keeps the files moved out of Humans.AuditLog byte-identical to their
+      originals.
+    -->
+    <Using Include="Microsoft.AspNetCore.Http" />
+    <Using Include="Microsoft.AspNetCore.Routing" />
+    <Using Include="Microsoft.Extensions.Logging" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <!--
+      No Humans.Infrastructure reference: Monitor owns no tables, so there is no DbContext, no
+      repository and no AddSectionDbContext call (G5-SECTION-TEMPLATE.md step 1). Scanner's
+      shape. DriveActivityMonitorService reads Google through GoogleIntegration's connector
+      abstraction and writes through IAuditLogService; neither needs Infrastructure.
+
+      Monitor is the *consumer* end of every edge it has — nothing in the repo references it
+      except DriveActivityMonitorJob, through Humans.Monitor.Contracts. That is what makes
+      Monitor -> GoogleIntegration + AuditLog legal where AuditLog -> GoogleIntegration was
+      not: AuditLog is a horizontal (peters-hard-rules.md), and Monitor is not.
+      MonitorArchitectureTests.NothingReferencesMonitor pins it.
+    -->
+    <ProjectReference Include="..\..\Humans.Interfaces\Humans.Interfaces.csproj" />
+    <ProjectReference Include="..\..\Humans.Domain\Humans.Domain.csproj" />
+    <ProjectReference Include="..\..\Humans.Application\Humans.Application.csproj" />
+    <ProjectReference Include="..\..\Humans.UI\Humans.UI.csproj" />
+    <ProjectReference Include="..\Humans.Monitor.Contracts\Humans.Monitor.Contracts.csproj" />
+    <ProjectReference Include="..\Humans.AuditLog.Contracts\Humans.AuditLog.Contracts.csproj" />
+    <ProjectReference Include="..\Humans.SystemSettings.Contracts\Humans.SystemSettings.Contracts.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <!--
+      No AspNetCore package here: those ship with no PackageVersion on purpose and would fail
+      the build with NU1010 under Central Package Management. Everything the controller needs
+      arrives through the framework reference Sdk.Razor already adds (design §15 step 1).
+      No EF Core either — the section owns no tables.
+    -->
+    <PackageReference Include="NodaTime" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <!-- The section owns its documentation (design §7a); keep it visible in Solution Explorer. -->
+    <None Include="**\*.md" />
+  </ItemGroup>
+
+</Project>
