+++ src/Sections/Humans.Monitor/Views/Monitor/GoogleSync.cshtml
@@ -1,6 +1,6 @@
 @using Humans.UI.Models.Tables
 @using Microsoft.AspNetCore.Html
-@model Humans.AuditLog.Models.GoogleSyncAuditListViewModel
+@model Humans.Monitor.Models.GoogleSyncAuditListViewModel
 @{
     ViewData["Title"] = Model.Title;
 }
