+++ src/Sections/Humans.Monitor/Views/Monitor/_ViewStart.cshtml
@@ -1,0 +1,3 @@
+@{
+    Layout = "_AdminLayout";
+}
